backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.12

# AI / LLM
langgraph==0.2.60
langchain-google-genai==2.0.7
langchain-core==0.3.28
google-generativeai==0.8.3

# Database
pymongo>=4.6,<5.0
motor==3.6.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# File Processing
Pillow==11.1.0
python-magic==0.4.27
PyMuPDF==1.24.14
pydub==0.25.1

# HTTP & Utils
httpx==0.28.1
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
aiofiles==24.1.0

# Serialization
orjson==3.10.12
